LayoutTests/platform/blackberry/fast/forms/input-search-should-have-white-background.html

<!DOCTYPE html>
<html>
<head>
<title>RIM Bug 1488</title>
<script src="../../../../fast/js/resources/js-test-pre.js"></script>
<script>
function runTest()
{
    if (window.layoutTestController)
        layoutTestController.dumpAsText();

    element = document.getElementById("target");
    shouldBe("document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color')", "'rgb(255, 255, 255)'");
}
</script>
</head>
<body onload="runTest()">
<p>This test can be used to verify that the background color of an &lt;input type=&quot;search&quot;&gt; is white.</p>
<div style="background-color: blue"><input id="target" type="search"/></div>
<pre id="console"></pre>
</body>
</html>
