LayoutTests/platform/blackberry/fast/forms/input-search-should-have-white-background-expected.txt

This test can be used to verify that the background color of an <input type="search"> is white.


PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(255, 255, 255)'
